+++ pages/tariff.vue
@@ -264,17 +264,19 @@ const loadData = async () => {
 
 const isDeleting = ref(false);
 const deleteRecord = async (id) => {
-  try {
-    isDeleting.value = true;
-    const res = await TariffService.delete(id);
-    list.value = list.value.filter((item) => item.id != id);
+  if (confirm("Are you sure to delete this record?")) {
+    try {
+      isDeleting.value = true;
+      const res = await TariffService.delete(id);
+      list.value = list.value.filter((item) => item.id != id);
 
-    serverErrors.value = {};
-    // handleReset();
-  } catch (error) {
-    serverErrors.value = error.errors;
-  } finally {
-    isDeleting.value = false;
+      serverErrors.value = {};
+      // handleReset();
+    } catch (error) {
+      serverErrors.value = error.errors;
+    } finally {
+      isDeleting.value = false;
+    }
   }
 };
 const record = reactive({
@@ -333,7 +335,7 @@ const updateRecord = async (id) => {
         item.name = record.name;
         item.start_date = record.startDate;
         item.end_date = record.endDate;
-        item.editMode = false
+        item.editMode = false;
         return item;
       }
       return item;
